+++ operator/internal/resources/aistore/target/statefulset.go
@@ -143,7 +143,7 @@ func NewInitContainerEnv(ais *aisv1.AIStore) (initEnv []corev1.EnvVar) {
 }
 
 func NewAISContainerEnv(ais *aisv1.AIStore) []corev1.EnvVar {
-	baseEnv := cmn.CommonEnv()
+	baseEnv := cmn.CommonEnv(ais)
 	if ais.Spec.HasGCPBackend() {
 		baseEnv = append(baseEnv, cmn.EnvFromValue(cmn.EnvGoogleCreds, filepath.Join(DefaultGCPDir, DefaultGCPConfig)))
 	}